pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quantpits"
version = "0.1.0"
description = "QuantPits: An advanced, production-ready quantitative trading system built on top of Microsoft Qlib."
authors = [
    {name = "DarkLink"}
]
requires-python = ">=3.8, <3.13"
dependencies = [
    "pyqlib",
    "pandas>=1.3.0",
    "numpy>=1.20.0,<2.0.0",
    "scipy>=1.7.0",
    "PyYAML>=5.4.1",
    "joblib>=1.0.1",
    "statsmodels>=0.13.0",
    "streamlit>=1.10.0",
    "plotly>=5.9.0"
]

[tool.setuptools.packages.find]
include = ["quantpits*"]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=quantpits --cov-report=term-missing"
testpaths = [
    "tests",
]
pythonpath = [
    "."
]
